Add support for homonymous entities
Your keychain may contain homonymous entities (many certificates with the same subject name) when you request a new certificate from Apple without deleting the previous one (because you don't want to regenerate all your provisioning profiles right now).

This commit allows to distinguish between homonymous entities by using the SHA-1 (which codesign accepts as an identity) of the certificate instead of its subject name. If you provide a SHA-1, it will be accepted. If you provide a subject name (like normal people do), it will be automatically substituted with the SHA-1 of the first matching certificate during the validation of the parameter.

Moreover, when displaying the list of available identities, homonymous certificates are grouped together and their SHA-1 are also displayed.