Allow using expired identities in resign
Due to [the expiration of WWDR Intermediate Certificate](https://developer.apple.com/support/certificates/expiration), signing identities may be incorrectly reported as invalid by the security tool, thus preventing the user from resigning the IPA without renewing the certificate.

While renewing the certificate is required for App Store submissions, in-house apps continue to work. Moreover, because renewing the certificate may require to revoke existing certificates, that renewal may not be possible in some cases (which is why this patch was written).

In order to remedy that issue, this patch provides the following changes:
 * We retrieve all signing identities.
 * Valid and expired (we cannot distinguish between expired certificate and expired CA) identities are kept. Other identities are discarded.
 * Valid identities take precedence over expired ones when searching for a certificate subject name (Working workflows should continue to work).
 * Expired identities are listed in white.
 * A warning is printed when the user select an expired identity.

In an ideal world, the signing identity would be extracted from the provisioning profile instead of asking the user. I may work on that in the more or less distant future. In the meantime, this patch should do the job.